website/requirements.txt
django-bulk-update==2.2.0
django-enumfields==0.10.1
django-extensions==2.0.7
django-polymorphic==2.0.2
django-redis==4.9.0
django-taggit==0.22.2
django-webpack-loader==0.6.0
Django==2.0.6
djangorestframework==3.8.2
gunicorn==19.8.1
ipdb==0.10.1
ipython==5.1.0
plumbum==1.6.6
psycopg2==2.6.2
pudb==2018.1
gitpython==2.1.10
pytest-django==3.1.2
python-memcached==1.59
PyYAML==3.12
Pyzotero==1.3.2
requests==2.18.4
wagtail==2.1
wagtailfontawesome==1.1.3
ipdb==0.10.1
